VM-Snapshot-Report.ps1, Version 1.0
Ready for merge if accepted.  PowerActions Report Script that reports on
VMs with snapshots along with their description, size of the snapshot in
GB, the VM's provisioned vs. used space in GB,
if the snapshot is the
current one being used, its parent snapshot (if there is one), and the
Power state of the VM itself. VM object is key (as it's the first
managed object in the output), enabling you the ability to right-click
an entry in the report to edit the target VM.