Updates to Set-HVPool
Updated Set-HVPool to address issue #132.  The -Start and -Stop parameter tested the wrong variable in the loop.  $item is undefined.  Furthermore, in general if the key/value pair specified $false as the value then the parameter wasn't seen in simple logic test such as if ($key -and $value) {}.  I've updated to use $PSBoundParameters instead.