Updates needed for v2.0
In v2.0 of the tool, the URI for Get-XVCMStatus changed from /api/ping to /api/status, updated line 15.  
In v2.0 of the tool, there were a number of changes necessary for New-XVCMRequest. Updated the Notes on Lines 145-173 to reflect new parameters, changed behavior and removed parameters. Added $opType on line 176 and 202 for the Operation type new to v2.0.  Removed line 181 and 196 for Source Cluster because it is no longer needed. Modified $DstCluster to accept null and removed the String Type definition to allow null to not be converted or the task creation will fail because it expects null as the assigned value.  Added $DstHost to line 184 and 199 to allow destination host input to override $DstCluster.